Added/Changed packages to separate entities from use case related files.
