+++ src/main/Java/ChangeEnvironmentUseCase/ChangeEnvironmentInputBoundary.java
@@ -1,4 +1,4 @@
-package mariogatchi;
+package ChangeEnvironmentUseCase;
 
 public interface ChangeEnvironmentInputBoundary{
    /*
